+++ uptimerobot/api/monitor.go
@@ -140,6 +140,7 @@ func (client UptimeRobotApiClient) CreateMonitor(req MonitorCreateRequest) (m Mo
 	data.Add("friendly_name", req.FriendlyName)
 	data.Add("url", req.URL)
 	data.Add("type", fmt.Sprintf("%d", monitorType[req.Type]))
+	data.Add("interval", fmt.Sprintf("%d", req.Interval))
 	switch req.Type {
 	case "port":
 		data.Add("sub_type", fmt.Sprintf("%d", monitorSubType[req.SubType]))
@@ -202,6 +203,7 @@ func (client UptimeRobotApiClient) UpdateMonitor(req MonitorUpdateRequest) (m Mo
 	data.Add("friendly_name", req.FriendlyName)
 	data.Add("url", req.URL)
 	data.Add("type", fmt.Sprintf("%d", monitorType[req.Type]))
+	data.Add("interval", fmt.Sprintf("%d", req.Interval))
 	switch req.Type {
 	case "port":
 		data.Add("sub_type", fmt.Sprintf("%d", monitorSubType[req.SubType]))