+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.2.0
+- resource monitor: add support for interval attribute (#7)
+- resource alert_contact: remove ForceNew on url (#12)
+- various documentation updates (#8, #10)
+
 # v0.1.1
 - Fix a bug in resourceMonitorUpdate (#5)
 - Fix a crash in resourceMonitorGet when the monitor doesn't exist (#6)